\name{checkconstraints}
\alias{checkconstraints}
\title{
Checking the possible/impossible household status under predefined constraints due to structural zeros.
}
\description{
Checking the possible/impossible household status under predefined constraints due to structural zeros.
}
\usage{
checkconstraints(data, neededpossiblehh, hh_size)
checkconstraints_format2(data, neededpossiblehh, hh_size)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data }{
  A household data matrix generated by calling samplinghouseholds.
  }
  \item{neededpossiblehh }{
  The number of possible households needed before checking is stopped.
  }
  \item{hh_size }{
  The household size for the households in \verb{data}.
  }
}
\details{
Given an input household data matrix, these functions will check the possible/impossible status of each household and also output the desired number of possible/impossible households separately. \verb{checkconstraints} checks constraints when the household head is included as an individual within the household while \verb{checkconstraints_format2} checks contraints when the household head is moved to the household level.

The predefined constraints/structural zeros already included are given below:
\itemize{
\item Valid Head: Each household must contain exactly one head and he/she must be at least 17 years old.
\item Valid Spouse: Each household cannot contain more than one spouse and he/she must be at least 17 years old. 
\item Valid Couple: No same sex couple and age difference between couples cannot exceed 52.
\item Valid Children: The household head must be older than the oldest child by at least 12. 
\item Valid Children-in-Law: The household head must be older than the oldest child-in-law by at least 10. 
\item Valid Parents: The youngest parent must be older than the household head by at least 13. 
\item Valid Parents-in-Law: The youngest parent-in-law must be older than the household head by at least 9. 
\item Valid Siblings: The age difference between the household head and siblings cannot exceed 33.
\item Valid Siblings-in-Law: The age difference between the household head and siblings-in-law cannot exceed 33.
\item Valid Grandchildren: The household head and spouse must both be at least 34 years old. Also, the household head must be older than the grandchildren by at least 30. 
}
}
\value{
A list containing information on checking result.
  \item{outcome }{An indicator vector for the possible/impossible household status under constraints.}
  \item{Households }{A data matrix for impossible households.}
  \item{Index }{A vector for the original indexes of households when possible households are found. Generally not to be used.}
  \item{synHouseholds }{A data matrix for possible households.}
  \item{possible }{The actual number of possible households returned.}
}

\author{
Quanli Wang
}

\keyword{constraints}
\keyword{impossible household}
\keyword{possible household}

