\name{households2individuals}
\alias{households2individuals}
\title{
Convert a household data matrix to the corresponding individual member data matrix.
}
\description{
Convert a household data matrix to the corresponding individual member data matrix.
}
\usage{
households2individuals(data, hh_size)
}
\arguments{
  \item{data}{
  Household data matrix.
}
  \item{hh_size}{
  The household size for the households in \verb{data}.
  }
}

\value{
Individual member data matrix.
}

\author{
Quanli Wang
}
\keyword{utility function}

