\name{initParameters}
\alias{initParameters}
\title{
MCMC model parameters initilization.
}
\description{
MCMC model parameters initilization.
}
\usage{
initParameters(data, hyper)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
  The input data to be used in the model.
}
  \item{hyper}{
  Hyper parameters for the prior distributions.
}
}

\value{
All the parameters that ae used and are to be updated by the model. Provide more details later once the individual parameters are documented.
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}

\author{
Quanli Wang
}
