\name{SampleEpsilonIndiv}
\alias{SampleEpsilonIndiv}

\title{
Michael: Edit Here
}
\description{
Michael: Edit Here
}
\usage{
SampleEpsilonIndiv(ErrorData)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ErrorData}{
Michael: Edit Here
}
}
\details{
Michael: Edit Here
}
\value{
Michael: Edit Here
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}

