\name{SampleTrueResponse}
\alias{SampleTrueResponse}
\title{
Michael: Edit here
}
\description{
Michael: Edit here
}
\usage{
SampleTrueResponse(ErrorData, orig, para, G_household, M, hyper)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ErrorData}{
Michael: Edit here
}
  \item{orig}{
Michael: Edit here
}
  \item{para}{
Michael: Edit here
}
  \item{G_household}{
Michael: Edit here
}
  \item{M}{
Michael: Edit here
}
  \item{hyper}{
Michael: Edit here
}
}
\details{
Michael: Edit here
}
\value{
Michael: Edit here
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
