\name{UpdateBeta}
\alias{UpdateBeta}
\title{
Update beta parameter
}
\description{
Update beta parameter.
}
\usage{
UpdateBeta(ba, bb, v)
}
\arguments{
  \item{ba}{
  hyper parameter a for beta.
}
  \item{bb}{
  hyper parameter b for beta.
}
  \item{v}{
  parameter v.
}
}

\value{
Updated beta parameter.
}

\author{
Quanli Wang
}

\keyword{ sampler }

