\name{checkconstraints}
\alias{checkconstraints}
\title{
Checking the possible/impossible household status under predefined constraints due to structure zeros.
}
\description{
Checking the possible/impossible household status under predefined constraints due to structure zeros.
}
\usage{
checkconstraints(data, neededpossiblehh)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data }{
  A household datmatrix generated by calling samplinghouseholds.
  }
  \item{neededpossiblehh }{
  The number of possible household one need before stopping checking.
  }
}
\details{
Given an input household data matrix, this functions will check the possible/impossible status of each household and also output the desired number of possible/impossible households separately.
}
\value{
A list containing information on checking result.
  \item{outcome }{An indicator vector for possible/impossible household status under constraints.}
  \item{Households }{A data matrix for impossible households}
  \item{Index }{A vector for the original indexes of households when possible households are found. Generally not to be used.}
  \item{synHouseholds }{A data matrix for possible households.}
  \item{possible }{The actural number of possible households returned.}
}

\author{
Quanli Wang
}

\keyword{constraints}
\keyword{impossible household}
\keyword{possible household}

