\name{households2individuals}
\alias{households2individuals}
\title{
Convert a household data matrix to the corresponding individual household member data matrix.
}
\description{
Convert a household data matrix to the corresponding individual household member data matrix.
}
\usage{
households2individuals(data)
}
\arguments{
  \item{data}{
  household data matrix.
}
}

\value{
individual household member data matrix.
}

\author{
Quanli Wang
}
\keyword{utility function}

