\name{initData}
\alias{initData}
\title{
Initilize the input data structure.
}
\description{
Initilize the input data structure.
}
\usage{
initData(household)
}
\arguments{
  \item{household}{
  A data matrix for the input household data.
}
}
\details{
Curently the variables and the level of each variable are harded coded. This needs to be changed later on once it is generalized.
}
\value{
A list object including all the necessary data variables needed by the  model.
}

\author{
Quanli Wang
}

