\name{samplehouseholds}
\alias{samplehouseholds}
\title{
Rcpp implementation for sampling household data without constraints.
}
\description{
Rcpp implementation for sampling household data without constraints.
}
\usage{
samplehouseholds(phi, w, pi, d, lambda, currrentbatch,
                nHouseholds, householdsize)
}
\arguments{
  \item{phi}{
  phi parameter
}
  \item{w}{
  w parameter
}
  \item{pi}{
  pi parameter
}
  \item{d}{
  the number of levels for each variable.
}
  \item{lambda}{
  A list of lambdas for household level variables.
}
  \item{currrentbatch}{
  The current batch number for the household data to be generated. The household ID will be generated based on this batch number.
}
  \item{nHouseholds}{
  The number of households to be generated by one call to this function.
}
  \item{householdsize}{
  The size of the household to be generated.
}
}
\details{
 This function allows the model to generate a batch of nHouseholds with each household of size householdsize. The generated household data will include both possible and impossible households.
}
\value{
A data matrix with each row for one household.
}

\author{
Quanli Wang
}
\keyword{sampler}
\keyword{household data without constraint}
