\name{UpdateAlpha}
\alias{UpdateAlpha}
\title{
Update alpha parameter
}
\description{
Update alpha parameter.
}
\usage{
UpdateAlpha(aa, ab, u)
}
\arguments{
  \item{aa}{
  hyper parameter a for alpha
}
  \item{ab}{
  hyper parameter b for alpha
}
  \item{u}{
  u parameter
}
}

\value{
updated alpha parameter.
}

\author{
Quanli Wang
}

\keyword{ sampler }

