\name{UpdatePi}
\alias{UpdatePi}
\title{
Update pi parameter
}
\description{
Update pi parameter.
}
\usage{
UpdatePi(alpha, z_HH_all, K)
}
\arguments{
  \item{alpha}{
  alpha parameter.
}
  \item{z_HH_all}{
  component indicator at the household level for all households, including impossible households.
}
  \item{K}{
  Maximum number of components at the household level.
}
}
\value{
updated pi parameter.
}

\author{
Quanli wang
}

\keyword{ sampler }

